G-7 - Unit tests pour Games DAO

See merge request dinf/projets/a20/eq05-gamesapi!5

+++ src/controllers/games.controller.go
@@ -68,7 +68,7 @@ func UpdateGame(c *gin.Context) {
 		c.JSON(gameErr.Status(), gameErr)
 		return
 	}
-	game.ID = uint(gameId)
+	game.ID = gameId
 	g, err := services.GamesService.UpdateGame(&game)
 	if errorUtils.IsEntityError(c, err) {
 		return
